try to test the types
