+++ setup.cfg
@@ -16,6 +16,9 @@ no_implicit_optional = False
 disallow_untyped_defs = False
 check_untyped_defs = True
 
+[mypy-tests.test_typing]
+strict = True
+
 [tool:isort]
 profile = black
 combine_as_imports = True
