+++ starlette/types.py
@@ -1,6 +1,6 @@
 import typing
 
-T = typing.TypeVar("T")
+AppType = typing.TypeVar("AppType")
 
 Scope = typing.MutableMapping[str, typing.Any]
 Message = typing.MutableMapping[str, typing.Any]
@@ -10,8 +10,8 @@ Send = typing.Callable[[Message], typing.Awaitable[None]]
 
 ASGIApp = typing.Callable[[Scope, Receive, Send], typing.Awaitable[None]]
 
-StatelessLifespan = typing.Callable[[T], typing.AsyncContextManager[None]]
+StatelessLifespan = typing.Callable[[AppType], typing.AsyncContextManager[None]]
 StatefulLifespan = typing.Callable[
-    [T], typing.AsyncContextManager[typing.Mapping[str, typing.Any]]
+    [AppType], typing.AsyncContextManager[typing.Mapping[str, typing.Any]]
 ]
-Lifespan = typing.Union[StatelessLifespan[T], StatefulLifespan[T]]
+Lifespan = typing.Union[StatelessLifespan[AppType], StatefulLifespan[AppType]]
